+++ ui/app/(creation-compte-widget)/espace-pro/widget/entreprise/fin/FinPage.tsx
@@ -1,0 +1,9 @@
+"use client"
+
+import { DepotRapideFin } from "@/app/(espace-pro)/_components/DepotRapideFin"
+
+export const WidgetEntrepriseMiseEnRelation = () => {
+  return <DepotRapideFin />
+}
+
+export default WidgetEntrepriseMiseEnRelation
