+++ ui/app/(creation-compte-widget)/espace-pro/widget/entreprise/offre/page.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next"
 import { DepotSimplifieCreationOffre } from "@/app/(espace-pro-creation-compte)/_components/DepotSimplifieCreationOffre"
 
 export const metadata: Metadata = {
-  title: "Formulaire de dépôt d'offre - compte créé - La bonne alternance",
+  title: "Formulaire de dépôt d'offre - Informations de l'offre - La bonne alternance",
 }
 
 export default function Page() {
