+++ ui/app/(espace-pro-creation-compte)/espace-pro/creation/fin/page.tsx
@@ -1,5 +1,3 @@
-"use client"
-
 import { DepotRapideFin } from "@/app/(espace-pro)/_components/DepotRapideFin"
 
 function Page() {
