+++ ui/app/test-widget/page.tsx
@@ -1,10 +1,15 @@
 import { fr } from "@codegouvfr/react-dsfr"
 import { Box } from "@mui/material"
 
+import type { Metadata } from "next"
 import { Footer } from "@/app/_components/Footer"
 import { PublicHeader } from "@/app/_components/PublicHeader"
 import { WidgetTester } from "@/app/_components/WidgetTester"
 import { getSession } from "@/utils/getSession"
+
+export const metadata: Metadata = {
+  title: "Formulaire de test des widgets - La bonne alternance",
+}
 
 export default async function Page() {
   const { user } = await getSession()
